Add support for QuickCheck-ing tests.
The syntax is as follows:

verifyUsing: quickcheck 10 100

means that we should use random testing to try to discover problems with
a emthod.

The first number specifies how many tests should pass.

The second number is optional, and it specifies the maximum number of
attempts.  The attempts may be more then the passed tests, because some
randomly generated inputs may not satisfy the pre-conditions on a method.
If the second number is omitted, then there is no limit on the number of
attempts, and so testing may take a long time.