Do not override existing SMTLIB files. (XXX: Avoid race)
We generate multiple theories, one for each aliasing configuration.
Previously, we simply over-wrote the old theory, so we only got the
last one.  This is clearly wrong, so now at least we pick a different
file for each.

The current solution is (hopefully!) only temporary, until we have a
way to name the different property configurations.