Redo support for SMTLIB2, by translating from SMTLIB1.
I am not too sure if this is a good idea, but it seems that things might
be a bit easier to maintain in this form.

Now the translations is like this:
HAJava -> SMTLIB1 -> SMTLIB2

Before, it used to be:
HAJava -> SMTLIB1
HAJava -> SMTLIB2

The two translations had a lot of duplication.

The drawback of the current implementation is that it more biased to
the older SMTLIB1 format.

Another possibility is to go: HAJava -> SMTLIB2 -> SMTLIB1 ...  Maybe one day.