Remove some dependencies of Verifier.Java.Common.
Ultimately, we'd like to remove all dependencies of this module on
abcBridge, because it interacts badly with the use of Template Haskell.
On Linux and Windows, the Template Haskell processor has trouble loading
our abcBridge library.

The main remaining work is to refactor things so LogMonad is either in a
different package or no longer used.