stopgap support for Stackage 1.5 LTS
Updated the build-sandbox.sh script to take advantage of Stackage
stuff. Unfortunately due to a bug in Cabal, we need to have copies of
the Stackage cabal.config scattered across the repo wherever we might
run a `cabal install` command. However we should now have no problems
with staging the tools and running the -t target in the same sandbox as
the regular build.

This is a stopgap only! Work continues on a proper Makefile that will
handle all of this much more efficiently and make for more reliable
test and build artifacts. That's just taking longer than I had expected,
so I'm getting the benefits of Stackage out to y'all sooner.