Implement CNF writer in JSS SAW backend.
Attempting to implement this four line function led to the previous 7
commits, a few future commits related to unsafe bit engine use in the
LSS and JSS SAW backends, and a few commits to the external packages
abcBridge and aig. Also, some future commits testing AIG and CNF
generation would be a good idea.