Fix 'fileAccess: does not exist (No such file or directory)' bug.
This bug has been around since I started working on SAW, but it was
usually only a minor annoyance. But last weekend it made some
long-running tests fail, so I decided to hunt it down. The problem was
that you can't call 'getPermissions' on a path that does not "exist",
and broken symlinks don't "exist".

Emacs creates broken symlinks for edited-but-not-saved buffers [1],
which was the typical reason I experienced this bug. The problem can
be reproduced by manually creating a broken symlink and running 'saw':

    ln -s non-existent-file broken-symlink
    saw

[1] http://stackoverflow.com/questions/5738170/why-does-emacs-create-temporary-symbolic-links-for-modified-files