Use a library to reimplement the code I fixed in my last commit.
The directory reading code had some weird `seq` calls in it -- perhaps
to avoid memory leaks? -- and seems to be equivalent to a few lines
using the 'directory-tree' library. The comments for the old
implementation mentioned it returned an *ordered* list of directory
contents. I don't expect that mattered, or know what "ordered" means
here, but path-based orderings could be recovered by adding a call to
'sort', e.g.

    recurseDirectory path = *sort .* toList . dirTree <$> build path